+++ DGIdGateway/YSFDefines.h
@@ -47,8 +47,8 @@ const unsigned char YSF_CM_INDIVIDUAL = 0x03U;
 const unsigned char YSF_MR_NOT_BUSY = 0x01U;
 const unsigned char YSF_MR_BUSY     = 0x02U;
 
-const unsigned int FCS_PORT = 62500U;
+const unsigned short FCS_PORT = 62500U;
 
-const unsigned int IMRS_PORT = 21110U;
+const unsigned short IMRS_PORT = 21110U;
 
 #endif
